Rename package to baseball_scraper
This is done so that a new package can be uploaded to pypi
